requirements.txt
google-generativeai>=0.8.5
trafilatura>=0.9.6
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=1.10.0
sentence-transformers>=2.2.2
numpy>=1.25.0
faiss-cpu>=1.7.4; platform_system!='Windows'
rapidfuzz>=3.0.0
typer[all]>=0.9.0
rich>=13.0.0
notion-client>=0.1.19
google-api-python-client>=2.80.0
google-auth>=2.21.0
google-auth-oauthlib>=1.0.0
ics>=0.7.2
pytest>=7.0.0
ruff>=0.12.0
black>=24.0.0
